+++ docs/db/operational_analysis.md
@@ -99,63 +99,3 @@
 | action           | String     | 操作类型    |
 | bk_process_ids   | NumberLong | 进程id    |
 | bk_process_names | String     | 进程名称    |
-
-## cc_ChartConfig
-
-#### 作用
-
-存放运营统计图表配置数据信息
-
-#### 表结构
-
-| 字段                  | 类型         | 描述       |
-|---------------------|------------|----------|
-| _id                 | ObjectId   | 数据唯一ID   |
-| config_id           | NumberLong | 图表配置数据ID |
-| report_type         | String     | 统计类型     |
-| name                | String     | 图表名称     |
-| bk_obj_id           | String     | 统计实例类型   |
-| width               | String     | 图表宽度     |
-| chart_type          | String     | 图表类型     |
-| field               | String     | 分类字段     |
-| x_axis_count        | NumberLong | x轴数量     |
-| create_time         | ISODate    | 创建时间     |
-| bk_supplier_account | String     | 开发商ID    |
-
-## cc_ChartData
-
-#### 作用
-
-存放运营统计数据
-
-#### 表结构
-
-| 字段                  | 类型       | 描述     |
-|---------------------|----------|--------|
-| _id                 | ObjectId | 数据唯一ID |
-| report_type         | String   | 统计类型   |
-| data                | Object   | 统计数据   |
-| create_time         | ISODate  | 创建时间   |
-| bk_supplier_account | String   | 开发商ID  |
-
-## cc_ChartPosition
-
-#### 作用
-
-存放运营统计中主机与其它模型实例的统计图表顺序数据
-
-#### 表结构
-
-| 字段                  | 类型         | 描述                  |
-|---------------------|------------|---------------------|
-| _id                 | ObjectId   | 数据唯一ID              |
-| bk_biz_id           | NumberLong | 业务id                |
-| position            | Object     | 图表顺序数据，数字小的图表对应位置靠前 |
-| bk_supplier_account | String     | 开发商ID               |
-
-#### position 字段结构
-
-| 字段   | 类型    | 描述       |
-|------|-------|----------|
-| host | Array | 主机图表顺序   |
-| inst | Array | 模型实例图表顺序 |
