+++ docs/support-file/config/templates/server#conf#common.yaml
@@ -21,15 +21,6 @@ esb:
   caFile:
   password:
 
-# operation_server专属配置
-operationServer:
-  timer:
-    # 00:00-23:59,运营统计定时收集数据时间点,默认是为00:30
-    #定时同步的时间点
-    spec: 15:30 # 00:00 - 23:59
-  # 禁用运营统计数据统计功能，默认false，如果设置为true，将无法查看定时统计的主机、模型实例等的变化数据
-  disableOperationStatistic: false
-
 #auth_server专属配置
 authServer:
   #蓝鲸权限中心地址,可配置多个,用,(逗号)分割
