--- resources/errors/common/default/operation_server.json
@@ -1,10 +1,0 @@
-{
-  "1116001": "获取业务、模块、主机数量失败",
-  "1116002": "新增统计报表失败",
-  "1116003": "统计报表已经存在",
-  "1116004": "删除统计图表失败",
-  "1116005": "获取统计图表失败",
-  "1116006": "更新统计图表失败",
-  "1116007": "获取图表数据失败",
-  "1116008": "更新图表位置失败"
-}
