--- resources/errors/common/default/statistics_server.json
@@ -1,9 +1,0 @@
-{
-  "1114001": "获取业务、模块、主机数量失败",
-  "1114002": "新增统计报表失败",
-  "1114003": "获取统计图表失败",
-  "1114004": "更新统计图表失败",
-  "1114005": "获取图表数据失败",
-  "1114006": "删除统计图表失败",
-  "1114007": "统计报表已经存在"
-}
