+++ resources/errors/common/default/topo_server.json
@@ -147,5 +147,6 @@
 	"1101126": "模型唯一校验(id: %d)和字段模板唯一校验(keys: %+v)冲突",
 	"1101127": "模板在模型(%s)应用时，会与业务(%d)下的自定义字段发生冲突。冲突的自定义字段：(bk_property_id: %s)",
 	"1101128": "该业务含有容器资源，禁止归档",
+	"1101129": "获取业务、模块、主机数量失败",
 	"": ""
 }
