--- src/scene_server/operation_server/service/init_func.go
@@ -1,32 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making 蓝鲸 available.
- * Copyright (C) 2017-2018 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * http://opensource.org/licenses/MIT
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package service
-
-import (
-	"net/http"
-
-	"configcenter/src/common"
-	httpheader "configcenter/src/common/http/header"
-)
-
-// InitFunc TODO
-func (o *OperationServer) InitFunc() {
-	header := make(http.Header, 0)
-	if "" == httpheader.GetSupplierAccount(header) {
-		httpheader.SetSupplierAccount(header, common.BKSuperOwnerID)
-		httpheader.SetUser(header, common.BKProcInstanceOpUser)
-	}
-
-	srvData := o.newSrvComm(header)
-	go srvData.lgc.TimerFreshData(srvData.ctx)
-}
