+++ src/source_controller/coreservice/service/translate.go
@@ -102,12 +102,6 @@ func (s *coreService) TranslateClassificationName(defLang language.DefaultCCLang
 		att.ClassificationID)
 }
 
-// TranslateOperationChartName TODO
-func (s *coreService) TranslateOperationChartName(defLang language.DefaultCCLanguageIf,
-	att metadata.ChartConfig) string {
-	return util.FirstNotEmptyString(defLang.Language("operation_chart_"+att.ReportType), att.Name, att.ReportType)
-}
-
 // TranslateAssociationType translate association type
 func (s *coreService) TranslateAssociationType(defLang language.DefaultCCLanguageIf,
 	assKind *metadata.AssociationKind) {
